Update wLedFont.h
I feel this is a better solution than the one assigned to lower-case t but didn't want to override someone else's work. I have the first 7 panels as zone0 and the 8th panel is zone 1. I use this character for displaying the time and numeric temperature value in zone0, then ºC in zone1.